+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - **New:** Rewrote display and data logic code.
 - **New:** Added `single` or `multiple` can be auto supported.
+- **New:** Added virtual scroll to support large data.
 - **New:** Added object `value` supported.
 - **New:** Added object for `v-model` supported.
 - **New:** Added `data` attributes of optgroup and option supported.